backend/requirements.txt
fastapi
uvicorn
python-multipart
timm
numpy
pillow
rasterio
opencv-python-headless
matplotlib 